pyproject.toml
[project]
name = "msme-chatbot-qa"
version = "0.1.0"
description = "MSME Chatbot QA with LangGraph Swarm Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # AI/ML Frameworks
    "langchain==0.3.26",
    "langchain-anthropic==0.3.9",
    "langchain-community==0.3.26",
    "langchain-core==0.3.76",
    "langchain-experimental==0.3.4",
    "langchain-google-genai==2.1.12",
    "langchain-mcp-adapters==0.1.7",
    "langchain-mongodb==0.6.2",
    "langchain-openai==0.3.28",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.5.4",
    "langgraph-checkpoint==2.1.1",
    "langgraph-checkpoint-mongodb==0.1.4",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk==0.1.74",
    "langgraph-supervisor==0.0.27",
    "langgraph-swarm==0.0.13",
    "langsmith==0.4.8",
    # OpenAI & Anthropic
    "openai==1.97.1",
    "anthropic==0.49.0",
    # Azure Services
    "azure-cognitiveservices-speech==1.43.0",
    # FastAPI & Web
    "fastapi==0.115.11",
    "uvicorn==0.34.0",
    "starlette==0.46.0",
    "python-multipart==0.0.20",
    "sse-starlette==2.3.6",
    # Streamlit
    "streamlit==1.44.1",
    "streamlit-mic-recorder==0.0.8",
    # Database
    "pymongo==4.11.3",
    "motor==3.7.0",
    "sqlalchemy==2.0.38",
    "psycopg2-binary==2.9.10",
    # Data Processing
    "pandas==2.2.3",
    "numpy==2.2.3",
    "pandasql==0.7.3",
    "pyarrow==19.0.1",
    # Visualization
    "matplotlib==3.10.1",
    "seaborn==0.13.2",
    "altair==5.5.0",
    "pydeck==0.9.1",
    # MCP Tools
    "mcp==1.16.0",
    "mcp-server==0.1.4",
    "cli-mcp==0.18.0",
    "fastmcp==2.12.4",
    # Utilities
    "python-dotenv==1.1.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.8.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "aiohttp==3.11.13",
    "requests==2.32.4",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "click==8.1.8",
    "typer==0.16.0",
    "rich==14.0.0",
    # Audio
    "pydub==0.25.1",
    "speechrecognition==3.14.2",
    # Flask
    "flask==3.1.0",
    "werkzeug==3.1.1",
    # Development Tools
    "ipython==9.0.1",
    "jinja2==3.1.6",
    "pyyaml==6.0.2",
    "toml==0.10.2",
    "gitpython==3.1.44",
    "tiktoken==0.9.0",
    "qdrant-client>=1.15.1",
    "langchain-qdrant>=0.2.1",
    "msme-chatbot-agents",
]

[tool.uv.sources]
msme-chatbot-agents = { path = "msme_chatbot_agents", editable = true }
